+++ WorldCities/WorldCitiesAPI/CHANGELOG.md
@@ -71,4 +71,5 @@ The following steps were used to generate this project:
 - Modify `appsettings.json` make sure to have JwtSetting:SecurityKey to contain a GUID or string with more than 256 bytes
 - Create `RegisterRequest.cs` contain the same as LoginRequest but with the added username
 - Create `RegisterResult.cs` similar to loginresult
-- Modify `AccountController.cs` add a new httppost for Register() 
+- Modify `AccountController.cs` add a new httppost for Register() and Id = Guid.NewGuid().ToString() to ApplicationUser
+- Modify `SeedController.cs` file add Id = Guid.NewGuid().ToString() to ApplicationUser